requirements.txt
bcrypt==4.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.1.1
Werkzeug==3.1.3
pandas==2.3.0
plotly==5.24.1
openai>=1.0.0
gunicorn==23.0.0
